Move STLC into tests
