+++ package.yaml
@@ -33,16 +33,6 @@ library:
   source-dirs: src
 
 tests:
-  examples:
-    main:                STLC.hs
-    source-dirs:         examples
-    ghc-options:
-    - -threaded
-    - -rtsopts
-    - -with-rtsopts=-N
-    dependencies:
-    - refinery
-
   refinery-test:
     main:                Spec.hs
     source-dirs:         test
